feat: additional changes to utilize the Metanthropes API
