+++ module/dice/metaeval.mjs
