+++ module/dice/metaexecute.mjs
