+++ module/dice/metaroll.mjs
