+++ module/dice/metarollextras.mjs
