+++ module/dice/metarollhandler.mjs
