+++ module/utils/dice-tools.mjs
@@ -1,0 +1,14 @@
+/**
+ *
+ * Helper function to extract the number of d10 dice from a given value
+ *
+ * @param {String} value
+ * @returns integer
+ */
+export async function metaExtractNumberOfDice(value) {
+	const match = value.match(/^(\d+)d10$/);
+	if (match) {
+		return parseInt(match[1], 10); // Convert the captured string to an integer
+	}
+	return null; // Return null if the string doesn't match the expected format
+}
